+++ module-service/src/main/java/com/mjc/school/service/impl/NewsService.java
@@ -5,13 +5,13 @@ import com.mjc.school.repository.model.impl.NewsModel;
 import com.mjc.school.service.BaseService;
 import com.mjc.school.service.dto.NewsDtoRequest;
 import com.mjc.school.service.dto.NewsDtoResponse;
+import com.mjc.school.service.exception.ResourceNotFoundException;
+import com.mjc.school.service.exception.ServiceExceptionMessage;
 import com.mjc.school.service.mapper.NewsMapper;
 import com.mjc.school.service.validation.Validate;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
@@ -26,16 +26,19 @@ public class NewsService implements BaseService<NewsDtoRequest, NewsDtoResponse,
     @Override
     public List<NewsDtoResponse> readAll() {
         //
-        return newsRepository.readAll().stream().map(newsMapper::newsToDto).collect(Collectors.toList());
+        return newsRepository.readAll().stream()
+                .map(newsMapper::newsToDto)
+                .collect(Collectors.toList());
     }
 
     @Override
     public NewsDtoResponse readById(Long id) {
         //
         Optional<NewsModel> newsModelOptional = newsRepository.readById(id);
-        if (newsModelOptional.isPresent()) {
-            return newsMapper.newsToDto(newsModelOptional.get());
-        } else throw new RuntimeException("No news with such id found");
+        return newsMapper.newsToDto(
+                newsModelOptional.orElseThrow(() ->
+                        new ResourceNotFoundException(String.format(ServiceExceptionMessage.NEWS_ID_DOES_NOT_EXIST.getErrorMessage(), id)))
+        );
     }
 
     @Override
@@ -43,10 +46,8 @@ public class NewsService implements BaseService<NewsDtoRequest, NewsDtoResponse,
     public NewsDtoResponse create(NewsDtoRequest createRequest) {
         //
         NewsModel newsModel = newsMapper.newsDtoToModel(createRequest);
-        newsModel.setCreateDate(LocalDateTime.now());
-        newsModel.setLastUpdateDate(LocalDateTime.now());
-        newsRepository.create(newsModel);
-        return newsMapper.newsToDto(newsModel);
+        NewsModel newNewsModel = newsRepository.create(newsModel);
+        return newsMapper.newsToDto(newNewsModel);
     }
 
     @Override
@@ -54,14 +55,21 @@ public class NewsService implements BaseService<NewsDtoRequest, NewsDtoResponse,
     public NewsDtoResponse update(NewsDtoRequest updateRequest) {
         //
         NewsModel updatedNews = newsMapper.newsDtoToModel(updateRequest);
-        updatedNews.setLastUpdateDate(LocalDateTime.now());
-        updatedNews.setCreateDate(newsRepository.readById(updatedNews.getId()).get().getCreateDate());
-        return newsMapper.newsToDto(newsRepository.update(updatedNews));
+        NewsModel newsModel = newsRepository.readById(updatedNews.getId())
+                .orElseThrow(() ->
+                        new ResourceNotFoundException(String.format(ServiceExceptionMessage.NEWS_ID_DOES_NOT_EXIST.getErrorMessage(), updatedNews.getId()))
+                );
+        newsModel.setTitle(updateRequest.getTitle());
+        newsModel.setContent(updatedNews.getContent());
+        return newsMapper.newsToDto(newsRepository.update(newsModel));
     }
 
     @Override
     public boolean deleteById(Long id) {
         //
+        if (!newsRepository.existById(id)) {
+            throw new ResourceNotFoundException(String.format(ServiceExceptionMessage.NEWS_ID_DOES_NOT_EXIST.getErrorMessage(), id));
+        }
         return newsRepository.deleteById(id);
     }
 }
